chore: put the library into Maintenance Mode (#28)
* chore: put the library into Maintenance Mode

This gem doesn't work with languages other than liquid, and was forced to be licensed GPLv3 because I forked the gem from another library when I created it. 

Now that [ayushn21/bridgetown-svg-inliner](https://github.com/ayushn21/bridgetown-svg-inliner) exists, which is MIT and works with the latest versions of Bridgetown, this library can go into maintenance mode.

The upgrade path is seamless, simply swap the gems out in your gemfile. 

If this breaks in any way, I will try to fix it but I will not be adding more features thanks to [ayushn21/bridgetown-svg-inliner](https://github.com/ayushn21/bridgetown-svg-inliner).